+++ CHANGES.rst
@@ -8,6 +8,9 @@ New Features:
 
 * docutils/Sphinx frontend: don't abort on encountering math/math_block, output
   the (LaTeX) math markup instead, along with printing a warning.
+* docutils frontend: raw inline text (with ``:format: 'rinoh'``) is parsed as
+  styled text
+
 
 Fixed: